make clean of rdf's user & article add some api about article update README.md

+++ migrate.sh
@@ -1,10 +1,17 @@
-#! /bin/bash
+#!/bin/bash
 
-MODEL_LIST="rdf users"
-for model in ${MODEL_LIST}; do
-    rm -rf ${model}/migrations/*
-    touch ${model}/migrations/__init__.py
-done
-rm -f db.sqlite3
-python3 manage.py makemigrations
-python3 manage.py migrate
+if [[ $1 == "clean" || $1 == "reset" ]]; then
+    MODEL_LIST="rdf users"
+    for model in ${MODEL_LIST}; do
+        rm -rf ${model}/migrations/*
+        touch ${model}/migrations/__init__.py
+    done
+    rm -f db.sqlite3
+
+    if [[ $1 == "reset" ]]; then
+        python3 manage.py makemigrations
+        python3 manage.py migrate
+    fi
+else
+    echo "Usage: bash migrate.sh <clean|reset>"
+fi
